requirements.txt
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
base58==2.0.1
bleach==3.2.1
blinker==1.4
boto3==1.16.13
botocore==1.19.13
Brotli==1.0.9
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
confuse==1.3.0
cycler==0.10.0
dash==1.17.0
dash-bootstrap-components==0.10.7
dash-colorscales==0.0.4
dash-core-components==1.13.0
dash-daq==0.5.0
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.0
decorator==4.4.2
defusedxml==0.6.0
docopt==0.6.2
dtale==1.22.0
entrypoints==0.3
enum-compat==0.0.3
Flask==1.1.2
Flask-Compress==1.8.0
flask-ngrok==0.0.25
future==0.18.2
gitdb==4.0.5
GitPython==3.1.11
htmlmin==0.1.12
idna==2.10
ImageHash==4.1.0
importlib-resources==3.3.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
kaleido==0.0.3.post1
kiwisolver==1.3.1
llvmlite==0.34.0
lxml==4.6.1
lz4==3.1.0
MarkupSafe==1.1.1
matplotlib==3.3.2
missingno==0.4.2
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.2
networkx==2.5
notebook==6.1.5
numba==0.51.2
numpy==1.19.4
packaging==20.4
pandas==1.1.4
pandas-profiling==2.9.0
pandocfilters==1.4.3
parso==0.7.1
pathtools==0.1.2
patsy==0.5.1
pexpect==4.8.0
phik==0.10.0
pickleshare==0.7.5
Pillow==8.0.1
pipreqs==0.4.10
plotly==4.12.0
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.13.0
ptyprocess==0.6.0
pyarrow==2.0.0
pycparser==2.20
pydeck==0.5.0
Pygments==2.7.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.2
requests==2.24.0
retrying==1.3.3
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.11.0
Send2Trash==1.5.0
six==1.15.0
smmap==3.0.4
squarify==0.4.3
statsmodels==0.12.1
streamlit==0.70.0
streamlit-pandas-profiling==0.0.6
strsimpy==0.1.9
sweetviz==1.1.2
tangled-up-in-unicode==0.0.6
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.51.0
traitlets==5.0.5
tzlocal==2.1
urllib3==1.25.11
validators==0.18.1
visions==0.5.0
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
xarray==0.16.1
yarg==0.1.9
